+++ workspaces/ui-v2/src/optic-components/diffs/render/AskForCommitMessage.tsx
@@ -58,7 +58,7 @@ export default function AskForCommitMessage(props: { hasChanges: boolean }) {
     });
     // If there are no batch commits (first commit) - link to the just created commit
     history.push(changelogPageRoute.linkTo(lastBatchCommitId || batchCommitId));
-    // TODO we need to trigger a new full page refresh - or trigger a refetch of data through out queries
+    // @nic TODO we need to trigger a new full page refresh - or trigger a refetch of data through out queries
   };
 
   return (
