+++ workspaces/ui-v2/src/optic-components/pages/diffs/ReviewEndpointDiffPage/index.ts
@@ -1,1 +1,1 @@
-export * from './ReviewEndpointDiffPage';
+export * from './ReviewEndpointDiffPageContainer';
